I think I know why it fails...
It wasn't getting the newest time upon each loop. Rather, if it always
thought it was the same time. So if it got it at 15:19, it thought it
was always 15:19. I've got it updating, but now I need to make it
realize it is the same time and move onto close_Win().